cleanup and fix locating for godog dependency
