+++ ast_test.go
@@ -1,162 +1,76 @@
 package godog
 
 import (
-	"bytes"
-	"go/format"
 	"go/parser"
 	"go/token"
-	"strings"
 	"testing"
 )
 
-var astMainFile = `package main
-
-import "fmt"
-
-func main() {
-	fmt.Println("hello")
-}`
-
-var astNormalFile = `package main
-
-import "fmt"
-
-func hello() {
-	fmt.Println("hello")
-}`
-
-var astTestMainFile = `package main
+var astContextSrc = `package main
 
 import (
-	"fmt"
-	"testing"
-	"os"
+	"github.com/DATA-DOG/godog"
 )
 
-func TestMain(m *testing.M) {
-	fmt.Println("hello")
-	os.Exit(0)
+func MyContext(s *godog.Suite) {
 }`
 
-var astPackAliases = `package main
+var astTwoContextSrc = `package lib
 
 import (
-	"testing"
-	a "fmt"
-	b "fmt"
+	"github.com/DATA-DOG/godog"
 )
 
-func TestMain(m *testing.M) {
-	a.Println("a")
-	b.Println("b")
+func ApiContext(s *godog.Suite) {
+}
+
+func DBContext(s *godog.Suite) {
 }`
 
-var astAnonymousImport = `package main
-
-import (
-	"testing"
-	_ "github.com/go-sql-driver/mysql"
-)
-
-func TestMain(m *testing.M) {
-}`
-
-var astLibrarySrc = `package lib
-
-import "fmt"
-
-func test() {
-	fmt.Println("hello")
-}`
-
-var astInternalPackageSrc = `package godog
-
-import "fmt"
-
-func test() {
-	fmt.Println("hello")
-}`
-
-func astProcess(src string, t *testing.T) string {
+func astContextParse(src string, t *testing.T) []string {
 	fset := token.NewFileSet()
 	f, err := parser.ParseFile(fset, "", []byte(src), 0)
 	if err != nil {
 		t.Fatalf("unexpected error while parsing ast: %v", err)
 	}
 
-	astDeleteTestMainFunc(f)
-
-	var buf bytes.Buffer
-	if err := format.Node(&buf, fset, f); err != nil {
-		t.Fatalf("failed to build source file: %v", err)
-	}
-
-	return buf.String()
+	return astContexts(f)
 }
 
-func TestShouldCleanTestMainFromSimpleTestFile(t *testing.T) {
-	actual := strings.TrimSpace(astProcess(astTestMainFile, t))
-	expect := `package main`
+func TestShouldGetSingleContextFromSource(t *testing.T) {
+	actual := astContextParse(astContextSrc, t)
+	expect := []string{"MyContext"}
 
-	if actual != expect {
-		t.Fatalf("expected output does not match: %s", actual)
+	if len(actual) != len(expect) {
+		t.Fatalf("number of found contexts do not match, expected %d, but got %d", len(expect), len(actual))
 	}
-}
 
-func TestShouldCleanTestMainFromFileWithPackageAliases(t *testing.T) {
-	actual := strings.TrimSpace(astProcess(astPackAliases, t))
-	expect := `package main`
-
-	if actual != expect {
-		t.Fatalf("expected output does not match: %s", actual)
+	for i, c := range expect {
+		if c != actual[i] {
+			t.Fatalf("expected context '%s' at pos %d, but got: '%s'", c, i, actual[i])
+		}
 	}
 }
 
-func TestShouldNotModifyNormalFile(t *testing.T) {
-	actual := strings.TrimSpace(astProcess(astNormalFile, t))
-	expect := astNormalFile
+func TestShouldGetTwoContextsFromSource(t *testing.T) {
+	actual := astContextParse(astTwoContextSrc, t)
+	expect := []string{"ApiContext", "DBContext"}
 
-	if actual != expect {
-		t.Fatalf("expected output does not match: %s", actual)
+	if len(actual) != len(expect) {
+		t.Fatalf("number of found contexts do not match, expected %d, but got %d", len(expect), len(actual))
+	}
+
+	for i, c := range expect {
+		if c != actual[i] {
+			t.Fatalf("expected context '%s' at pos %d, but got: '%s'", c, i, actual[i])
+		}
 	}
 }
 
-func TestShouldNotModifyMainFile(t *testing.T) {
-	actual := strings.TrimSpace(astProcess(astMainFile, t))
-	expect := astMainFile
+func TestShouldNotFindAnyContextsInEmptyFile(t *testing.T) {
+	actual := astContextParse(`package main`, t)
 
-	if actual != expect {
-		t.Fatalf("expected output does not match: %s", actual)
-	}
-}
-
-func TestShouldMaintainAnonymousImport(t *testing.T) {
-	actual := strings.TrimSpace(astProcess(astAnonymousImport, t))
-	expect := `package main
-
-import (
-	_ "github.com/go-sql-driver/mysql"
-)`
-
-	if actual != expect {
-		t.Fatalf("expected output does not match: %s", actual)
-	}
-}
-
-func TestShouldNotModifyLibraryPackageSource(t *testing.T) {
-	actual := strings.TrimSpace(astProcess(astLibrarySrc, t))
-	expect := astLibrarySrc
-
-	if actual != expect {
-		t.Fatalf("expected output does not match: %s", actual)
-	}
-}
-
-func TestShouldNotModifyGodogPackageSource(t *testing.T) {
-	actual := strings.TrimSpace(astProcess(astInternalPackageSrc, t))
-	expect := astInternalPackageSrc
-
-	if actual != expect {
-		t.Fatalf("expected output does not match: %s", actual)
+	if len(actual) != 0 {
+		t.Fatalf("expected no contexts to be found, but there was some: %v", actual)
 	}
 }
